修复一些小bug并引入第三方微信登录功能

+++ gulimall-product/src/main/java/com/xunqi/gulimall/product/service/impl/CategoryServiceImpl.java
@@ -8,6 +8,7 @@ import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.xunqi.common.utils.PageUtils;
 import com.xunqi.common.utils.Query;
 import com.xunqi.gulimall.product.dao.CategoryDao;
+import com.xunqi.gulimall.product.entity.CategoryBrandRelationEntity;
 import com.xunqi.gulimall.product.entity.CategoryEntity;
 import com.xunqi.gulimall.product.service.CategoryBrandRelationService;
 import com.xunqi.gulimall.product.service.CategoryService;
@@ -84,8 +85,15 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryDao, CategoryEntity
     @Override
     public void removeMenuByIds(List<Long> asList) {
         //TODO 检查当前删除的菜单，是否被别的地方引用
-        //逻辑删除
-        baseMapper.deleteBatchIds(asList);
+        List<CategoryBrandRelationEntity> categoryBrandRelation =
+                categoryBrandRelationService.list(new QueryWrapper<CategoryBrandRelationEntity>().in("catelog_id", asList));
+
+        if (categoryBrandRelation.size() == 0) {
+            //逻辑删除
+            baseMapper.deleteBatchIds(asList);
+        } else {
+            throw new RuntimeException("该菜单下面还有属性，无法删除!");
+        }
     }
